// src/app/models/sp.ts
export interface Sp {
  sp_primno?: number;
  sp_srk_no?: number;
  sp_bcmno?: number;
  sp_no1?: number;
  sp_no2?: number;
  sp_frm_kod?: string;
  sp_frm_ksad?: string;
  sp_frmd_kod?: string;
  sp_sk_kod?: string;
  sp_referans?: string;
  sp_siptrh?: any | null;
  sp_bitis?: string;
  sp_sde_no?: number;
  sp_st_kod?: string;
  sp_yuktrh?: any | null;
  sp_gsip_yil?: number;
  sp_gsip_no?: number;
  sp_per_no?: number;
  sp_veren?: string;
  sp_aciklama?: string;
  uk?: string;
  updt?: Date | null;
  iuk?: string;
  idt?: Date | null;
  sp_onay?: string;
  sp_dept_no?: number;
  sp_label_kod?: string;
  sp_liste?: number;
  sp_onaytrh?: Date | null;
  sp_per_no2?: number;
  sp_ode_plan?: string;
  sp_teslim?: string;
  sp_onay2?: number;
  sp_ipt_ned?: string;
  sp_ipt_gs_id?: number;
  sp_ihrc_tip?: string;
  sp_afrm_id?: number;
  sp_akom?: number;
  sp_banh_id?: number;
  sp_svkfrm_id?: number;
  sp_rev?: number;
  sp_dosya?: string;
  sp_fad_id?: number;
  sp_asp_id?: number;
  sp_asp_bcmno?: number;
  sp_asp_no1?: number;
  sp_asp_no2?: number;
  sp_akt_trh?: Date | null;
  sp_nak_tut?: number;
  sp_sig_tut?: number;
  sp_takip_no?: string;
  sp_onaytrh2?: Date | null;
  sp_aciklama1?: string;
  sp_aciklama2?: string;
  sp_svk_adres?: string;
  sp_stop?: number;
  sp_ei_id?: number;
  sp_hspdvz_kod?: string;
  sp_islt_id?: number;
  sp_islt_kod?: string;
  sp_anlasma_turu?: string;
  sp_urtt_id?: number;
  sp_urtt_kod?: string;
  sp_vade?: number;
  sp_onay_aciklama?: string;
  sp_muh_cari_kod?: string;
  sp_ode_tur?: number;
  sp_ulke_kod?: string;
  sp_sevk_adres?: string;
  sp_veren2?: string;
  sp_cikis_trh?: any | null;

  Where?: any;
}

export interface Spgrp {
  spg_primno: number;
  spg_ad: string;
}